Time: 191 ms (94.68%), Space: 195.5 MB (95.53%) - LeetHub

+++ 3259-maximum-energy-boost-from-two-drinks/3259-maximum-energy-boost-from-two-drinks.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long maxEnergyBoost(vector<int>& energyDrinkA, vector<int>& energyDrinkB) {
         long long a = 0, b = 0;
-        for (size_t i = 0; i < energyDrinkA.size(); ++i) {
+        for (long long i = 0; i < energyDrinkA.size(); ++i) {
             long long newA = max(a + energyDrinkA[i], b);
             long long newB = max(b + energyDrinkB[i], a);
             a = newA;
